Handle receiving strings from UART
Handles receiving strings from UART and processing them into command
strings in the main UART shell task.

When an RX interrupt event occurs, the character is read. If this is a
carriage return character, notify that a complete command has been
received. Otherwise, try to put the character in a ringbuffer. If the
ringbuffer is full, notify that an error has occurred. Process these
notifications in the UART shell task.